# docs/swagger.yaml
basePath: /
host: localhost:9090
info:
  contact:
    email: ooooooooooos@163.com
    name: silenceLamb
    url: http://www.swagger.io/support
  title: 快递代取 API
  version: 1.0.1
paths:
  /express/create:
    post:
      description: 这是一个新增订单接口
      parameters:
      - description: 快递单号
        in: formData
        name: code
        required: true
        type: string
      - description: 收货地址
        in: formData
        name: address
        required: true
        type: string
      - description: 收货日期
        in: formData
        name: receive_date
        required: true
        type: string
      - description: 订单费用
        in: formData
        name: price
        required: true
        type: string
      - description: 取件码
        in: formData
        name: receive_code
        required: true
        type: string
      - description: 创建人 id
        in: formData
        name: create_id
        required: true
        type: string
      - description: 创建人头像
        in: formData
        name: create_img
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 新增订单
      tags:
      - 快递订单
  /express/info:
    get:
      description: 这是一个订单详情接口
      parameters:
      - description: 订单 id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 订单详情
      tags:
      - 快递订单
  /express/list:
    get:
      description: 这是一个订单列表接口
      parameters:
      - description: 页码
        in: query
        name: page
        required: true
        type: string
      - description: 条数
        in: query
        name: size
        required: true
        type: string
      - description: 接单状态
        in: query
        name: status
        type: integer
      - description: 接单人
        in: query
        name: receiver_id
        type: string
      - description: 创建人
        in: query
        name: create_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 订单列表
      tags:
      - 快递订单
  /upload:
    post:
      description: 上传文件接口
      parameters:
      - description: 上传的文件
        in: formData
        name: file
        required: true
        type: string
      produces:
      - multipart/form-data
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 上传文件
      tags:
      - 上传文件
  /user/info:
    get:
      description: 用户详情接口
      parameters:
      - description: identity
        in: query
        name: identity
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 用户详情
      tags:
      - 用户
  /user/login:
    post:
      description: 用户登录接口
      parameters:
      - description: code
        in: formData
        name: code
        required: true
        type: string
      - description: 名字
        in: formData
        name: name
        required: true
        type: string
      - description: 头像
        in: formData
        name: avatarUrl
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 用户登录
      tags:
      - 用户
swagger: "2.0"